OG-122: split lookup window
added configuration item `relayLookupWindowParts` (which defaults to
"1"), to split the lookup window (defined in blocks) to multiple
requests.
added a code to increase the lookupWindowParts in case the lookup
transaction fails on "query returned more than XXX events"
(note that the increase stays in memory. if the client reloads, it will
use the configured value again)